Experiment: nicer go syntax highlight
Try and get code blocks and lists highlighted nicely and also directive
comments like `//go:build`, `//go:generate`

This was a bit of a struggle since matching anything inside `/* */`
comments is pretty much the same as matching anywhere in code (i.e. it's
not really possible to tell if a given line is part of the comments or
not). This is an issue with any syntax groups that have `contains=ALL`
or `contains=ALLBUT` so I had to overwrite these by copying them from
the system config file into here and then manually appending the custom
value