Experiment with a new way to trim Git branches
Where 'trim' means clean up old branches that I no longer need. The
current script (`./bin/git-trim`) works fine in repos that use merge
commits, but not so well in repos which rebase changes from pull
requests. This new script is an attempt to work in both cases.

Also set `git fetch` to also prune because that's more useful in this
case, since the new script uses existence of remote references to
decide if a branch needs to be trimmed