gpg: explicitly set `pinentry-program`
This program changed for me after an update from the `curses` one I'm
used to to a GTK one. I believe due to a change in the packaging[1] and
the fact my system apparently has all the libs needed to support
`pinentry-gtk` in all the right places.

Plus, being explicit is just nicer.

[1] https://gitlab.archlinux.org/archlinux/packaging/packages/pinentry/-/commit/88ba81630b8536b5e905c2ba342e3d54efd9041c