Fix i3 termination
Well, this was some fun learning, the key point is that `i3` will disown
programs run via `exec` (i.e. `setsid(1)` on them). I think it makes
some effort to terminate all applications with a window class, but for
things I start without a window (e.g. `mozillavpn`) they will be left
running. So previously, on start

* `i3` starts, `stderr` connected to a pipe
* `logger` starts, stdin connected to that same pipe
* `i3` spawns some children, they inherit file descriptors, i.e. with
  that pipe attached to their `stderr`

Then on termination of `i3` there are still some children left holding
the pipe's file descriptor, so `logger` will not terminate (since the
pipe is not closed), per docs[1] this means:

> If the pipeline is not executed asynchronously..., the shell waits
for all commands in the pipeline to complete.

So `i3 | logger` will keep running as long as those (now orphaned) child
processes are left running.

With this change: `logger` is run 'asynchronously' (per docs[2]) and
so[3]

> This is known as executing the command in the background, and these
are referred to as asynchronous commands. The shell does not wait for
the command to finish, and the return status is 0 (true).

So `i3 > >(logger)` will exit when `i3` exits: `logger` may still be
left running as long as there are orphaned processes from `i3` left
around, but it will mean `xinit` will stop.

Link: https://www.gnu.org/software/bash/manual/bash.html#Pipelines [1]
Link: https://www.gnu.org/software/bash/manual/bash.html#Process-Substitution [2]
Link: https://www.gnu.org/software/bash/manual/bash.html#Lists [3]