Fix up GPG pinentry handling
This is a followup to 47bb7566a9d6717c24a795fa0d95fc93069acd5f:
unsetting `PINENTRY_USER_DATA` in interactive shells just means
`gpg-conf` uses whatever it had in its environment at start time (likely
`gtk` since interactive programs are the first to try and query this at
login). I _could_ add an explicit start of the `gpg-conf` from a non-X
environment, but instead just be explicit about the setting of this
variable in interactive bash sessions.

To be specific about the problem with the previous setup, `gpg(1)` says:

> When calling the gpg-agent component gpg sends a set of environment
variables to gpg-agent.  The names  of  these  variables  can  be listed
using the command:

    gpg-connect-agent 'getinfo std_env_names' /bye | awk '$1=="D" {print $2}'

The output of which includes `PINENTRY_USER_DATA`. So if we unset this,
`gpg-agent` either doesn't receive this or just ignores it